+++ testBalloon-framework-core/src/jsHostedMain/kotlin/de/infix/testBalloon/framework/core/internal/integration/KotlinJsTestIntegration.kt
@@ -10,6 +10,7 @@ import de.infix.testBalloon.framework.core.internal.integration.TestSessionRelay
 import de.infix.testBalloon.framework.core.internal.logError
 import de.infix.testBalloon.framework.core.internal.logInfo
 import de.infix.testBalloon.framework.core.testPlatform
+import de.infix.testBalloon.framework.shared.internal.ReportingMode
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
@@ -40,8 +41,17 @@ internal fun TestElement.registerWithKotlinJsTestFramework() {
                     }
                 }
             } else {
+                val elementName =
+                    if (isTopLevelSuite || TestSession.global.reportingMode == ReportingMode.INTELLIJ_IDEA) {
+                        // A qualified path name for suites ensures proper nesting display in IntelliJ IDEA, but
+                        // duplicates path elements in XML and HTML file reports.
+                        testElementPath.fullyQualifiedReportingName
+                    } else {
+                        // Simple element names below the top level work for file reports.
+                        testElementPath.elementReportingName
+                    }
                 kotlinJsTestFramework.suite(
-                    testElementPath.modeDependentReportingName,
+                    elementName,
                     ignored = !testElementIsEnabled
                 ) {
                     testElementChildren.forEach {
@@ -52,7 +62,7 @@ internal fun TestElement.registerWithKotlinJsTestFramework() {
         }
 
         is Test -> {
-            kotlinJsTestFramework.test(testElementPath.modeDependentReportingName, ignored = !testElementIsEnabled) {
+            kotlinJsTestFramework.test(testElementPath.elementReportingName, ignored = !testElementIsEnabled) {
                 TestSessionRelay.resultReceivingPromise(this)
             }
         }