+++ testBalloon-framework-shared/src/commonMain/kotlin/de/infix/testBalloon/framework/shared/internal/Constants.kt
@@ -18,7 +18,11 @@ public object Constants {
     public const val JVM_DISCOVERY_RESULT_PROPERTY: String = "testFrameworkDiscoveryResult"
     public const val JVM_DISCOVERY_RESULT_PROPERTY_GETTER: String = "getTestFrameworkDiscoveryResult"
 
-    public const val JUNIT_ENGINE_NAME: String = "de.infix.testBalloon"
+    public const val JUNIT4_ENTRY_POINT_SIMPLE_CLASS_NAME: String = "TestBalloonJUnit4"
+    public const val JUNIT4_RUNNER_CLASS_NAME: String =
+        "$CORE_INTERNAL_PACKAGE_NAME.integration.TestBalloonJUnit4Runner"
+
+    public const val JUNIT_ENGINE_ID: String = "de.infix.testBalloon"
 
     public const val GRADLE_EXTENSION_NAME: String = "testBalloon"
     public const val GRADLE_PROPERTY_PREFIX: String = "testBalloon"